+++ codex/views/opds/auth.py
@@ -7,6 +7,7 @@ from rest_framework.authentication import (
 
 from codex.authentication import BearerTokenAuthentication
 from codex.views.auth import AuthMixin
+from codex.views.opds.user_agent import get_user_agent_name
 
 
 class OPDSAuthMixin(AuthMixin):
@@ -17,3 +18,10 @@ class OPDSAuthMixin(AuthMixin):
         BearerTokenAuthentication,
         SessionAuthentication,
     )
+
+    @property
+    def user_agent_name(self) -> str:
+        """Memoize user agent name."""
+        if self._user_agent_name is None:  # pyright: ignore[reportUnnecessaryComparison]
+            self._user_agent_name = get_user_agent_name(self.request)  # pyright: ignore[reportAttributeAccessIssue,reportUninitializedInstanceVariable], # ty: ignore[unresolved-attribute]
+        return self._user_agent_name